audio to cc finished, started with organ setup pane

+++ src/Controller/MixerSubController.h
@@ -18,8 +18,8 @@ class MixerSubController : public SubController
     void SelectNextSource(int channelIndex);
     void SwapVolumeOverride(int channelIndex);
 
-	 void SetChannelLevelLeft(bool master, int channelIndex, int value);
-    void SetChannelLevelRight(bool master, int channelIndex, int value);
+	 void SetChannelLevelLeft(bool master, int channelIndex, double value);
+    void SetChannelLevelRight(bool master, int channelIndex, double value);
     void SetChannelGateLeft(bool master, int channelIndex, bool value);
     void SetChannelGateRight(bool master, int channelIndex, bool value);
